Reduce Authentication factory API
