+++ core/src/test/java/com/jaspersoft/android/sdk/service/auth/AuthenticationServiceTest.java
@@ -1,0 +1,45 @@
+package com.jaspersoft.android.sdk.service.auth;
+
+import org.junit.Before;
+import org.junit.Rule;
+import org.junit.Test;
+import org.junit.rules.ExpectedException;
+import org.mockito.Mock;
+import org.mockito.MockitoAnnotations;
+
+import static org.mockito.Matchers.any;
+import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.verifyNoMoreInteractions;
+import static org.mockito.Mockito.when;
+
+/**
+ * @author Tom Koptel
+ */
+public class AuthenticationServiceTest {
+
+    @Mock
+    AuthPolicy mAuthPolicy;
+    @Mock
+    Credentials mCredentials;
+
+    private AuthenticationService authenticatorUnderTest;
+
+    @Rule
+    public ExpectedException mExpectedException = ExpectedException.none();
+
+    @Before
+    public void setupMocks() throws Exception {
+        MockitoAnnotations.initMocks(this);
+        when(mCredentials.applyPolicy(any(AuthPolicy.class))).thenReturn("cookie");
+        authenticatorUnderTest = new AuthenticationService(mAuthPolicy);
+    }
+
+    @Test
+    public void testAuthenticate() throws Exception {
+        authenticatorUnderTest.authenticate(mCredentials);
+
+        verify(mCredentials).applyPolicy(mAuthPolicy);
+        verifyNoMoreInteractions(mCredentials);
+        verifyNoMoreInteractions(mAuthPolicy);
+    }
+}
